+++ src/point3.rs
@@ -4,10 +4,10 @@ use num_traits::MulAdd;
 /// A 3D point.
 pub type Point3 = Vec3;
 
-impl MulAdd<f64, Vec3> for Point3 {
+impl MulAdd<f32, Vec3> for Point3 {
     type Output = Point3;
 
-    fn mul_add(self, a: f64, b: Vec3) -> Self::Output {
+    fn mul_add(self, a: f32, b: Vec3) -> Self::Output {
         // self * a + b
         Point3::new(
             self.e[0].mul_add(a, b.e[0]),
