+++ src/random.rs
@@ -1,11 +1,11 @@
 use rand::{distributions::Uniform, prelude::*, thread_rng};
 
 pub struct Random {
-    distribution: Uniform<f64>,
+    distribution: Uniform<f32>,
 }
 
 impl Random {
-    pub fn sample(&self) -> f64 {
+    pub fn sample(&self) -> f32 {
         let mut rng = thread_rng();
         self.distribution.sample(&mut rng)
     }
@@ -14,7 +14,7 @@ impl Random {
 impl Default for Random {
     fn default() -> Self {
         Self {
-            distribution: Uniform::from(0.0f64..1.0),
+            distribution: Uniform::from(0.0f32..1.0),
         }
     }
 }
